Abstracted chart legend
